pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "skskills"
version = "0.1.0"
description = "Sovereign Agent Skills — MCP-native skill framework replacing OpenClaw"
readme = "README.md"
license = {text = "GPL-3.0-or-later"}
requires-python = ">=3.10"
authors = [
    {name = "smilinTux", email = "admin@smilintux.org"},
]
keywords = [
    "ai", "agent", "skills", "mcp", "sovereign",
    "knowledge", "tools", "hooks", "decentralized",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: GNU General Public License v3 or later (GPLv3+)",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Software Development :: Libraries",
]

dependencies = [
    "click>=8.1",
    "mcp>=1.0",
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "rich>=13.0",
]

[project.optional-dependencies]
capauth = [
    "capauth>=0.1.0",
]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "pytest-asyncio>=0.23",
    "black>=24.0",
    "ruff>=0.3",
]

[project.scripts]
skskills = "skskills.cli:main"

[project.urls]
Homepage = "https://skills.smilintux.org"
Repository = "https://github.com/smilinTux/skskills"

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 99
target-version = ["py310"]

[tool.ruff]
line-length = 99
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
addopts = "-v --tb=short"
asyncio_mode = "auto"
